@@ -1,41 +1,42 @@
-﻿namespace Services.Order.Domain
+﻿namespace Services.Order.Domain.AggregateModels.Order;
+
+using Services.Order.Domain.Events;
+using Services.Order.Domain.SeedWork;
+using System;
+using System.Collections.Generic;
+
+public class Order : Entity, IAggregateRoot
 {
-    using Services.Order.Domain.SeedWork;
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
+    List<OrderItem> orderItems;
+    Address adress;
+    Guid buyerId;
+    Guid paymentMethodId;
+    OrderStatus status;
+    string description;
+    DateTime created;
 
-    public class Order : Entity, IAggregateRoot
+    public Order(Guid id)
     {
-        private List<OrderItem> _orderItems;
-        private Guid _customerId;
+        this.id = id;
+        status = OrderStatus.Submitted;
+        orderItems = new List<OrderItem>();
+        created = DateTime.UtcNow;
+    }
 
-        public Order(Guid id, Guid customerId)
-        {
-            Id = id;
-            _customerId = customerId;
-            _orderItems = new List<OrderItem>();
-        }
+    public Order(
+        Guid id,
+        Guid userId,
+        string username,
+        Card card) : this(id)
+    {
+        AddOrderStartedDomainEvent(userId, username, card);
+    }
 
-        public Guid CustomerId
-        {
-            get => _customerId;
-            private set { _customerId = value; }
-        }
-
-        public IReadOnlyCollection<OrderItem> OrderItems => _orderItems;
-
-        public void AddOrderItem(OrderItem orderItem)
-        {
-            var existingOrderItem = _orderItems.FirstOrDefault(x => x.Name == orderItem.Name);
-
-            if (existingOrderItem != null)
-            {
-                existingOrderItem.AddQuantity(orderItem.Quantity);
-                return;
-            }
-
-            _orderItems.Add(orderItem);
-        }
+    void AddOrderStartedDomainEvent(
+        Guid userId,
+        string username,
+        Card card)
+    {
+        AddDomainEvent(new OrderStartedDomainEvent(this, userId, username, card));
     }
 }
