@@ -1,28 +1,46 @@
-﻿namespace Services.Order.Domain
+﻿namespace Services.Order.Domain.AggregateModels.Order;
+
+using Services.Order.Domain.Exceptions;
+using Services.Order.Domain.SeedWork;
+using System;
+
+public class OrderItem : Entity
 {
-    using Services.Order.Domain.SeedWork;
-    using System;
+    string productName;
+    decimal unitPrice;
+    int units;
 
-    public class OrderItem : Entity
+    public string ProductName { get => productName; }
+    public decimal UnitPrice { get => unitPrice; }
+    public int Units { get => units; }
+
+    public OrderItem(Guid id, string productName, decimal unitPrice, int units)
     {
-        private Guid _orderId;
-        private string _name;
-        private int _quantity;
+        if (id == default)
+            throw new OrderingDomainException("invalid id");
 
-        public OrderItem(Guid orderId, string name, int quantity)
-        {
-            _orderId = orderId;
-            _name = name;
-            _quantity = quantity;
-        }
+        if (string.IsNullOrEmpty(productName))
+            throw new OrderingDomainException("product name cannot be empty");
 
-        public Guid OrderId => _orderId;
-        public string Name => _name;
-        public int Quantity => _quantity;
+        if (unitPrice < 1)
+            throw new OrderingDomainException($"Unit price cannot be lower than 1. Unit price was {unitPrice}");
 
-        public void AddQuantity(int value)
-        {
-            _quantity += value;
-        }
+        if (units < 0)
+            throw new OrderingDomainException($"Units cannot be less than 0. Units was {units}");
+
+        this.id = id;
+        this.productName = productName;
+        this.unitPrice = unitPrice;
+        this.units = units;
     }
+
+    public void AddUnits(int units)
+    {
+        if (units > 0)
+            throw new OrderingDomainException("Invalid units");
+
+        this.units += units;
+    }
+
+   
 }
