@@ -1,93 +1,85 @@
-﻿namespace Services.Order.Domain.SeedWork
+﻿namespace Services.Order.Domain.SeedWork;
+
+using MediatR;
+using System;
+using System.Collections.Generic;
+
+public abstract class Entity
 {
-    using System;
-    using MediatR;
-    using System.Collections.Generic;
+    int? requestedHashCode;
+    protected Guid id;
 
-    public abstract class Entity
+    public virtual Guid Id
     {
-        int? _requestedHashCode;
-        Guid _id;
-        
-        public virtual  Guid Id 
+        get => id;
+    }
+
+    List<INotification> _domainEvents;
+    public IReadOnlyCollection<INotification> DomainEvents => _domainEvents?.AsReadOnly();
+
+    public void AddDomainEvent(INotification eventItem)
+    {
+        _domainEvents = _domainEvents ?? new List<INotification>();
+        _domainEvents.Add(eventItem);
+    }
+
+    public void RemoveDomainEvent(INotification eventItem)
+    {
+        _domainEvents?.Remove(eventItem);
+    }
+
+    public void ClearDomainEvents()
+    {
+        _domainEvents?.Clear();
+    }
+
+    public bool IsTransient()
+    {
+        return id == default;
+    }
+
+    public override bool Equals(object obj)
+    {
+        if (obj == null || !(obj is Entity))
+            return false;
+
+        if (Object.ReferenceEquals(this, obj))
+            return true;
+
+        if (this.GetType() != obj.GetType())
+            return false;
+
+        Entity item = (Entity)obj;
+
+        if (item.IsTransient() || this.IsTransient())
+            return false;
+        else
+            return item.Id == id;
+    }
+
+    public override int GetHashCode()
+    {
+        if (!IsTransient())
         {
-            get
-            {
-                return _id;
-            }
-            protected set
-            {
-                _id = value;
-            }
+            if (!requestedHashCode.HasValue)
+                requestedHashCode = id.GetHashCode() ^ 31;
+
+            return requestedHashCode.Value;
         }
+        else
+            return base.GetHashCode();
 
-        private List<INotification> _domainEvents;
-        public IReadOnlyCollection<INotification> DomainEvents => _domainEvents?.AsReadOnly();
+    }
+    public static bool operator ==(Entity left, Entity right)
+    {
+        if (Object.Equals(left, null))
+            return (Object.Equals(right, null)) ? true : false;
+        else
+            return left.Equals(right);
+    }
 
-        public void AddDomainEvent(INotification eventItem)
-        {
-            _domainEvents = _domainEvents ?? new List<INotification>();
-            _domainEvents.Add(eventItem);
-        }
-
-        public void RemoveDomainEvent(INotification eventItem)
-        {
-            _domainEvents?.Remove(eventItem);
-        }
-
-        public void ClearDomainEvents()
-        {
-            _domainEvents?.Clear();
-        }
-
-        public bool IsTransient()
-        {
-            return this.Id == default(Guid);
-        }
-
-        public override bool Equals(object obj)
-        {
-            if (obj == null || !(obj is Entity))
-                return false;
-
-            if (Object.ReferenceEquals(this, obj))
-                return true;
-
-            if (this.GetType() != obj.GetType())
-                return false;
-
-            Entity item = (Entity)obj;
-
-            if (item.IsTransient() || this.IsTransient())
-                return false;
-            else
-                return item.Id == this.Id;
-        }
-
-        public override int GetHashCode()
-        {
-            if (!IsTransient())
-            {
-                if (!_requestedHashCode.HasValue)
-                    _requestedHashCode = this.Id.GetHashCode() ^ 31; // XOR for random distribution (http://blogs.msdn.com/b/ericlippert/archive/2011/02/28/guidelines-and-rules-for-gethashcode.aspx)
-
-                return _requestedHashCode.Value;
-            }
-            else
-                return base.GetHashCode();
-
-        }
-        public static bool operator ==(Entity left, Entity right)
-        {
-            if (Object.Equals(left, null))
-                return (Object.Equals(right, null)) ? true : false;
-            else
-                return left.Equals(right);
-        }
-
-        public static bool operator !=(Entity left, Entity right)
-        {
-            return !(left == right);
-        }
+    public static bool operator !=(Entity left, Entity right)
+    {
+        return !(left == right);
     }
 }
