@@ -1,7 +1,6 @@
-﻿namespace Services.Order.Domain.SeedWork
+﻿namespace Services.Order.Domain.SeedWork;
+
+public interface IRepository<T> where T : IAggregateRoot
 {
-    public interface IRepository<T> where T : IAggregateRoot
-    {
-        IUnitOfWork UnitOfWork { get; }
-    }
+    IUnitOfWork UnitOfWork { get; }
 }
