@@ -1,12 +1,11 @@
-﻿namespace Services.Order.Domain.SeedWork
-{
-    using System;
-    using System.Threading;
-    using System.Threading.Tasks;
+﻿namespace Services.Order.Domain.SeedWork;
 
-    public interface IUnitOfWork : IDisposable
-    {        
-        Task<int> SaveChangesAsync(CancellationToken cancellationToken = default(CancellationToken));
-        Task<bool> SaveEntitiesAsync(CancellationToken cancellationToken = default(CancellationToken));
-    }
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+public interface IUnitOfWork : IDisposable
+{
+    Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
+    Task<bool> SaveEntitiesAsync(CancellationToken cancellationToken = default);
 }
