@@ -1,48 +1,47 @@
-﻿namespace Services.Order.Domain.SeedWork
+﻿namespace Services.Order.Domain.SeedWork;
+
+using System.Collections.Generic;
+using System.Linq;
+
+public abstract class ValueObject
 {
-    using System.Collections.Generic;
-    using System.Linq;
-
-    public abstract class ValueObject
+    protected static bool EqualOperator(ValueObject left, ValueObject right)
     {
-        protected static bool EqualOperator(ValueObject left, ValueObject right)
+        if (ReferenceEquals(left, null) ^ ReferenceEquals(right, null))
         {
-            if (ReferenceEquals(left, null) ^ ReferenceEquals(right, null))
-            {
-                return false;
-            }
-            return ReferenceEquals(left, null) || left.Equals(right);
+            return false;
+        }
+        return ReferenceEquals(left, null) || left.Equals(right);
+    }
+
+    protected static bool NotEqualOperator(ValueObject left, ValueObject right)
+    {
+        return !(EqualOperator(left, right));
+    }
+
+    protected abstract IEnumerable<object> GetEqualityComponents();
+
+    public override bool Equals(object obj)
+    {
+        if (obj == null || obj.GetType() != GetType())
+        {
+            return false;
         }
 
-        protected static bool NotEqualOperator(ValueObject left, ValueObject right)
-        {
-            return !(EqualOperator(left, right));
-        }
+        var other = (ValueObject)obj;
 
-        protected abstract IEnumerable<object> GetEqualityComponents();
+        return this.GetEqualityComponents().SequenceEqual(other.GetEqualityComponents());
+    }
 
-        public override bool Equals(object obj)
-        {
-            if (obj == null || obj.GetType() != GetType())
-            {
-                return false;
-            }
+    public override int GetHashCode()
+    {
+        return GetEqualityComponents()
+            .Select(x => x != null ? x.GetHashCode() : 0)
+            .Aggregate((x, y) => x ^ y);
+    }
 
-            var other = (ValueObject)obj;
-
-            return this.GetEqualityComponents().SequenceEqual(other.GetEqualityComponents());
-        }
-
-        public override int GetHashCode()
-        {
-            return GetEqualityComponents()
-             .Select(x => x != null ? x.GetHashCode() : 0)
-             .Aggregate((x, y) => x ^ y);
-        }
-
-        public ValueObject GetCopy()
-        {
-            return this.MemberwiseClone() as ValueObject;
-        }
+    public ValueObject GetCopy()
+    {
+        return this.MemberwiseClone() as ValueObject;
     }
 }
